IGNITE-22745 Re-factor C++ code
