+++ modules/platforms/python/cpp_module/utils.cpp
@@ -1,0 +1,70 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements. See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+#include "utils.h"
+
+#include <ignite/odbc/diagnostic/diagnosable.h>
+#include <ignite/common/detail/defer.h>
+
+#include <Python.h>
+
+PyObject* py_get_class(const char* module_name, const char* class_name) {
+    auto pyignite3_mod = PyImport_ImportModule(module_name);
+
+    if (!pyignite3_mod)
+        return nullptr;
+
+    auto conn_class = PyObject_GetAttrString(pyignite3_mod, class_name);
+    Py_DECREF(pyignite3_mod);
+
+    return conn_class;
+}
+
+bool check_errors(ignite::diagnosable& diag) {
+    auto &records = diag.get_diagnostic_records();
+    if (records.is_successful())
+        return true;
+
+    std::string err_msg;
+    switch (records.get_return_code()) {
+        case SQL_INVALID_HANDLE:
+            err_msg = "Invalid object handle";
+            break;
+
+        case SQL_NO_DATA:
+            err_msg = "No data available";
+            break;
+
+        case SQL_ERROR:
+            auto record = records.get_status_record(1);
+            err_msg = record.get_message_text();
+            break;
+    }
+
+    // TODO: IGNITE-22226 Set a proper error here, not a standard one.
+    PyErr_SetString(PyExc_RuntimeError, err_msg.c_str());
+
+    return false;
+}
+
+const char* py_object_get_typename(PyObject* obj) {
+    if (!obj || !obj->ob_type || !obj->ob_type->tp_name) {
+        return "Unknown";
+    }
+
+    return obj->ob_type->tp_name;
+}
